pyproject.toml
[tool.poetry]
name = "database-replication"
version = "0.1.0"
description = ""
authors = ["joaoallmeida <joaoallmeida96@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
ipykernel = "^6.29.5"
sqlescapy = "^1.0.1"
requests = "^2.32.3"
psycopg2-binary = "^2.9.9"
pyarrow = "^17.0.0"
polars = "^1.5.0"
unidecode = "^1.3.8"
sqlalchemy = "^2.0.32"
pandas = "^2.2.2"
dbt-core = "^1.8.6"
dbt-postgres = "^1.8.2"
pyyaml = "^6.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
